Road-surface internal geometry now uses RoadVec2 / RoadVec3 backed by glam::DVec2/DVec3 at backend.rs (line 8). Long-lived span, node, terrain-clip, earthwork, cache, query, and debug internals were migrated off Godot vectors. Remaining Vector2 / Vector3 usage is limited to allowed boundaries: graph/API input, GDExtension conversion, render upload, debug output, and Cavalier’s non-Godot Vector2 alias. Cleaned old RoadVec-as-Vector test aliases in the internal surface test modules. Preserved the spec nuance around push_full_roadbed_contour: it still exists only as pre-ownership rail/footprint input, not as a rendered fallback carrier.

+++ rust/src/simulation/network/surface/earthwork.rs
@@ -81,10 +81,10 @@ impl RoadSurfaceSystem {
         for &chunk in chunks {
             let (chunk_min, chunk_max) = self.chunk_bounds(chunk);
             terrain.reset_visual_region_from_source_world(
-                chunk_min.x,
-                chunk_min.z,
-                chunk_max.x,
-                chunk_max.z,
+                chunk_min.x as f32,
+                chunk_min.z as f32,
+                chunk_max.x as f32,
+                chunk_max.z as f32,
             );
 
             let Some(entry) = self.earthwork_chunk_cache.get(&chunk) else {
